package/libgtk3: fix build failure on libinput due to the presence of atk-bridge-2.0 in gtk+-3.0.pc
Patch 0001-disable-atk-bridge.patch misses to remove:
atk_pkgs += ['atk-bridge-2.0']
that is not affected by any meson variable, this results in having libgtk3
to have atk-bridge-2.0 in its gtk+-3.0.pc and so other libraries(i.e.
libinput) tries to {pkgconf}find it failing. So let's convert the patch to
git format and add the removal of such line.

Fixes:
still not found

Signed-off-by: Giulio Benetti <giulio.benetti@benettiengineering.com>